+++ engine/docs/changelog.md
@@ -1,4 +1,11 @@
 # TBD
+### Features
+* Updated the destroy-enclave process adding `destroy-files-artifact-expansion-volumes` flow on it
+
+### Breaking Changes
+* Added new `kurtosisBackend` argument in `EnclaveManager.NewEnclaveManager`
+  * Users must pass an `backend_interface.KurtosisBackend` argument in all `NewEnclaveManager` calls
+
 ### Changes
 * Upgraded to `container-engine-lib` 0.14.0, which implement `files artifact expansion volume` and `files artifact expander` objects